feat(company): upgrade buildings

+++ building/service.go
@@ -7,14 +7,15 @@ import (
 
 type (
 	Building struct {
-		Id              uint64              `db:"id" json:"id"`
-		Name            string              `db:"name" json:"name"`
-		WagesHour       uint64              `db:"wages_per_hour" json:"wages_per_hour"`
-		AdminHour       uint64              `db:"admin_per_hour" json:"admin_per_hour"`
-		MaintenanceHour uint64              `db:"maintenance_per_hour" json:"maintenance_per_hour"`
-		Downtime        *uint8              `db:"downtime" json:"downtime"`
-		Requirements    []*resource.Item    `json:"requirements"`
-		Resources       []*BuildingResource `json:"resources"`
+		Id              uint64  `db:"id" json:"id"`
+		Name            string  `db:"name" json:"name"`
+		WagesHour       uint64  `db:"wages_per_hour" json:"wages_per_hour"`
+		AdminHour       uint64  `db:"admin_per_hour" json:"admin_per_hour"`
+		MaintenanceHour uint64  `db:"maintenance_per_hour" json:"maintenance_per_hour"`
+		Downtime        *uint16 `db:"downtime" json:"downtime"`
+
+		Requirements []*resource.Item    `json:"requirements"`
+		Resources    []*BuildingResource `json:"resources"`
 	}
 
 	BuildingResource struct {
